kidswebapp/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
PyMongo==4.7.3
pymongo[srv]==4.7.3
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
Werkzeug==3.0.3
dnspython==2.6.1
certifi
